Registering to the dynamic dispatcher moved to the hpp
